Don't run stress_sending_fill for all capacities
The problem is the wrap around of the receiver position with capacities
that are **not** a power of two.

For example taking a capacity of 5. The maximum receiver position is 63
(6 bits), which is index 3. The next value will be 0 (with wrap around),
which is index 0. This means we've skipped index 4.

The crate doesn't guaranteed FIFO ordering so it's fine, but this test
assumes it.