Fix data race when dropping oneshot::Sender/Receiver
This is similar to commit 3c232a17da84fc538f5de1856bd1f5eb274d685c, both
the problem and the manner in which it's solved.

In the Drop implementation for oneshot::Sender it would wake the
receiver after indicating it no longer held access, by unsetting the
SENDER_ALIVE bit in the shared status. This leads to a data
race/use-after-free if the Receiver is dropped at the same time and
drops the Shared data, which the Sender then accesses to wake the
Receiver.

The fix is the same as done in commit
3c232a17da84fc538f5de1856bd1f5eb274d685c, by adding a bit indicating the
Sender still has access to the shared data. This increases the number of
atomic fetch_and from 1 to 2. See commit
3c232a17da84fc538f5de1856bd1f5eb274d685c for additional details.