Check access bit in oneshot::Receiver::try_reset
Before it would if the SENDER_ALIVE bit was not set. This meant that the
Sender could still be alive, as the SENDER_ACCESS bit could still be
set. This causes a problem if the overwriting of the status (in
try_reset) was done before the original Sender unset the SENDER_ACCESS
bit, leading to double-free and use-after-free.