Add Receiver::(try_)peek
Allows the user to peek at the next available message in the inbox. The
API is similar to that of the try_recv and recv methods, but return &T
(instead of T).

Both methods need mutable access for peeking because of the waker
situation, which only supports a single waker for the receiver. At the
moment I don't think it's worth it to make the reiceiver more complex
simply to support immutable reference for peeking.